khety-backend/khety-ml/requirements.txt
flask==3.1.3
flask-cors==6.0.2
h5py==3.14.0
keras==3.12.1
numpy==2.2.6
pillow==12.2.0
requests==2.32.5
tensorflow==2.21.0
werkzeug==3.1.8
